feat: Final implementation of the frontend with a backend using JSON-server.
The implementation of the backend using a JSON server is for demonstration purposes only and has been completed. The JSON server is not ideal for storing large amounts of data, so the db.json file may be cleared after some usage. A better backend will be set up in the future.